+++ packages/cli/src/cli/cmd/show/files.ts
@@ -9,8 +9,8 @@ import { resolveOverriddenLocale } from "@lingo.dev/_spec";
 export default new Command()
   .command("files")
   .description("Print out the list of files managed by Lingo.dev")
-  .option("--source", "Only show source files")
-  .option("--target", "Only show target files")
+  .option("--source", "Only show source files, files containing the original translations")
+  .option("--target", "Only show target files, files containing translated content")
   .helpOption("-h, --help", "Show help")
   .action(async (type) => {
     const ora = Ora();
@@ -27,7 +27,7 @@ export default new Command()
 
         const buckets = getBuckets(i18nConfig);
         for (const bucket of buckets) {
-          for (const bucketConfig of bucket.config) {
+          for (const bucketConfig of bucket.paths) {
             const sourceLocale = resolveOverriddenLocale(i18nConfig.locale.source, bucketConfig.delimiter);
             const sourcePath = bucketConfig.pathPattern.replace(/\[locale\]/g, sourceLocale);
             const targetPaths = i18nConfig.locale.targets.map((_targetLocale) => {